Extracted `REDUNDANT_SEMICOLON` rule check from logic associated with `WRONG_NEWLINES` rule
### What's done:
- `REDUNDANT_SEMICOLON` rule check moved to earlier checks to remove unnecessary semicolons before executing the rest of the rules.
- Moved several warning and fix tests related to `REDUNDANT_SEMICOLON` rule.

Closes #1863, #1783
