+++ diktat-rules/src/test/resources/test/paragraph3/newlines/OneLineFunctionExpected.kt
@@ -2,5 +2,4 @@ package test.paragraph3.newlines
 
 class Example {
     fun doubleA(): Int = 2 * a
-    fun doubleA(): Int = 2 * a
 }
