+++ diktat-rules/src/test/resources/test/paragraph3/newlines/OneLineFunctionTest.kt
@@ -2,9 +2,6 @@ package test.paragraph3.newlines
 
 class Example {
     fun doubleA(): Int {
-        return 2 * a;
-    }
-    fun doubleA(): Int {
         return 2 * a
     }
 }
