+++ diktat-rules/src/test/resources/test/paragraph3/newlines/SizeParameterListExpected.kt
