+++ diktat-rules/src/test/resources/test/paragraph3/newlines/SizeParameterListTest.kt
