+++ diktat-rules/src/test/resources/test/paragraph3/semicolons/SemicolonsExpected.kt
